Add tests and some changes

+++ src/main/java/com/example/springbootguide/controller/DepartmentController.java
@@ -1,12 +1,11 @@
 package com.example.springbootguide.controller;
 
 import com.example.springbootguide.DTO.DepartmentDTO;
+import com.example.springbootguide.exception.RequestValidationException;
 import com.example.springbootguide.service.DepartmentService;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -26,18 +25,29 @@ public class DepartmentController {
     }
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public DepartmentDTO createDepartment(@RequestBody @Valid DepartmentDTO departmentDTO) {
         return departmentService.createDepartment(departmentDTO);
     }
 
     @DeleteMapping("/{departmentId}")
-    public void deleteDepartment(@PathVariable("departmentId") @Positive(message = "Id must be greater than 0") @NotNull Long id) {
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteDepartment(@PathVariable("departmentId") Long id) {
+        if (id == null || id <= 0) {
+            throw new RequestValidationException("Id must be greater than 0 and can't be empty");
+        }
         departmentService.deleteDepartment(id);
     }
 
     @PutMapping("/{departmentId}")
-    public void updateDepartment(@PathVariable("departmentId") @Positive(message = "Id must be greater than 0") @NotNull Long id,
-                                 @RequestParam(value = "name", required = false) @NotBlank String name) {
+    public void updateDepartment(@PathVariable("departmentId") Long id,
+                                 @RequestParam(value = "departmentName", required = false) String name) {
+        if (id == null || id <= 0) {
+            throw new RequestValidationException("Id must be greater than 0 and can't be empty");
+        }
+        if (name == null || name.isEmpty()) {
+            throw new RequestValidationException("Name can't be empty");
+        }
         departmentService.updateDepartment(id, name);
     }
 }
